Crudely slice things up into multiple files
